+++ native/SampleApps/AuthFlowTester/AuthFlowTesterUITests/PageObjects/LoginPageObject.swift
@@ -97,76 +97,19 @@ class LoginPageObject {
     ) -> Void {
         tap(settingsButton())
         tap(loginOptionsButton())
-        setSwitchField(useWebServerFlowSwitch(), value: useWebServerFlow)
-        setSwitchField(useHybridSwitch(), value: useHybridFlow)
-        
-        if let staticAppConfig = staticAppConfig {
-            let configJSON = buildConfigJSON(consumerKey: staticAppConfig.consumerKey, redirectUri: staticAppConfig.redirectUri, scopes: staticScopes)
-            importConfig(configJSON, isStaticConfiguration: true)
-        }
-        tap(saveStaticConfigButton())
-        
-        if let dynamicAppConfig = dynamicAppConfig {
-            let configJSON = buildConfigJSON(consumerKey: dynamicAppConfig.consumerKey, redirectUri: dynamicAppConfig.redirectUri, scopes: dynamicScopes)
-            importConfig(configJSON, isStaticConfiguration: false)
-            tap(saveDynamicConfigButton())
-        }
-        
-        let discoveryResultJSON = buildDiscoveryResultJSON(loginHost: discoveryLoginHost, username: discoveryUsername)
-        importDiscoveryResult(discoveryResultJSON)
-        tap(saveSimulatedResultButton())
-        
-        tap(loginOptionsCloseButton())
-    }
-    
-    private func buildConfigJSON(consumerKey: String, redirectUri: String, scopes: String) -> String {
-        let config: [String: String] = [
-            BootConfigJSONKeys.consumerKey: consumerKey,
-            BootConfigJSONKeys.redirectUri: redirectUri,
-            BootConfigJSONKeys.scopes: scopes
-        ]
-        guard let jsonData = try? JSONSerialization.data(withJSONObject: config, options: []),
-              let jsonString = String(data: jsonData, encoding: .utf8) else {
-            return "{}"
-        }
-        return jsonString
+        let loginOptionsPage = LoginOptionsPageObject(testApp: app)
+        loginOptionsPage.configure(
+            staticAppConfig: staticAppConfig,
+            staticScopes: staticScopes,
+            dynamicAppConfig: dynamicAppConfig,
+            dynamicScopes: dynamicScopes,
+            useWebServerFlow: useWebServerFlow,
+            useHybridFlow: useHybridFlow,
+            discoveryLoginHost: discoveryLoginHost,
+            discoveryUsername: discoveryUsername
+        )
     }
 
-    private func buildDiscoveryResultJSON(loginHost: String, username: String) -> String {
-        let config: [String: String] = [
-            "login_hint": username,
-            "my_domain": loginHost
-        ]
-        guard let jsonData = try? JSONSerialization.data(withJSONObject: config, options: []),
-              let jsonString = String(data: jsonData, encoding: .utf8) else {
-            return "{}"
-        }
-        return jsonString
-    }
-    
-    private func importConfig(_ jsonString: String, isStaticConfiguration: Bool = true) {
-        tap(importConfigButton(useStaticConfiguration: isStaticConfiguration))
-        
-        // Wait for alert to appear
-        let alert = importConfigAlert()
-        _ = alert.waitForExistence(timeout: timeout)
-        
-        // Type into the alert's text field
-        let textField = importConfigTextField()
-        textField.typeText(jsonString)
-        
-        tap(importAlertButton())
-    }
-
-    private func importDiscoveryResult(_ jsonString: String) {
-        tap(importDiscoveryResultButton())
-        let alert = importDiscoveryResultAlert()
-        _ = alert.waitForExistence(timeout: timeout)
-        let textField = alert.textFields.firstMatch
-        textField.typeText(jsonString)
-        tap(importDiscoveryResultAlertImportButton())
-    }
-    
     // MARK: - UI Element Accessors
     
     private func loginNavigationBar() -> XCUIElement {
@@ -238,80 +181,7 @@ class LoginPageObject {
     private func toolbarDoneButton() -> XCUIElement {
         return app.toolbars["Toolbar"].buttons["Done"]
     }
-    
-    private func useWebServerFlowSwitch() -> XCUIElement {
-        return app.switches["Use Web Server Flow"]
-    }
 
-    private func useHybridSwitch() -> XCUIElement {
-        return app.switches["Use Hybrid Flow"]
-    }
-
-    private func staticConfigurationSection() -> XCUIElement {
-        return app.buttons["Static Configuration"]
-    }
-    
-    private func consumerKeyField() -> XCUIElement {
-        return app.textFields["consumerKeyTextField"]
-    }
-    
-    private func callbackUrlField() -> XCUIElement {
-        return app.textFields["callbackUrlTextField"]
-    }
-    
-    private func scopesField() -> XCUIElement {
-        return app.textFields["scopesTextField"]
-    }
-    
-    private func saveStaticConfigButton() -> XCUIElement {
-        return app.buttons["Save static config"]
-    }
-
-    private func saveDynamicConfigButton() -> XCUIElement {
-        return app.buttons["Save dynamic config"]
-    }
-    
-    /// Returns the import button for either the static or dynamic configuration section.
-    /// The LoginOptionsView has two BootConfigEditor sections - the first for static config, the second for dynamic config.
-    private func importConfigButton(useStaticConfiguration: Bool = true) -> XCUIElement {
-        let buttons = app.buttons.matching(identifier: "importConfigButton")
-        let index = useStaticConfiguration ? 0 : 1
-        return buttons.element(boundBy: index)
-    }
-    
-    private func importConfigAlert() -> XCUIElement {
-        return app.alerts["Import Configuration"]
-    }
-    
-    private func importConfigTextField() -> XCUIElement {
-        // Access text field through the alert - SwiftUI alert TextFields are accessed this way
-        return importConfigAlert().textFields.firstMatch
-    }
-    
-    private func importAlertButton() -> XCUIElement {
-        return importConfigAlert().buttons["Import"]
-    }
-
-    private func importDiscoveryResultButton() -> XCUIElement {
-        return app.buttons["importDiscoveryResultButton"]
-    }
-
-    private func importDiscoveryResultAlert() -> XCUIElement {
-        return app.alerts["Import Discovery Result"]
-    }
-
-    private func importDiscoveryResultAlertImportButton() -> XCUIElement {
-        return importDiscoveryResultAlert().buttons["Import"]
-    }
-
-    private func saveSimulatedResultButton() -> XCUIElement {
-        return app.buttons["saveSimulatedResultButton"]
-    }
-
-    private func loginOptionsCloseButton() -> XCUIElement {
-        return app.buttons["loginOptionsCloseButton"]
-    }
-    
     private func advancedAuthCloseButton() -> XCUIElement {
         return app.otherElements["TopBrowserBar"].buttons["Close"]
     }
@@ -349,18 +219,6 @@ class LoginPageObject {
         
         textField.typeText(value)
         tapIfPresent(toolbarDoneButton())
-    }
-    
-    private func setSwitchField(_ switchField: XCUIElement, value: Bool) {
-        _ = switchField.waitForExistence(timeout: timeout)
-        
-        // Switch values are "0" (off) or "1" (on) in XCTest
-        let currentValue = (switchField.value as? String) == "1"
-        
-        // Only tap if the current state differs from desired state
-        if currentValue != value {
-            tap(switchField)
-        }
     }
     
     // MARK: - Other
